GUI: Fix opening .ts files in Watch mode on Windows
Gstreamer does not come with a decoder for mp2 audio which is common in .ts files but relies on codecs from either ffmpeg or mpg123.

We already enabled gst-libav to enable using codecs from ffmpeg. Problem is that the default ffmpeg build config of gvsbuild disables all codecs except some.

Fix is to enable building mp2float decoder in FFmpeg so it is available to Gstreamer allowing opening/watching files as expected.