Fix crash on files with incorrect timestamps (err: Application provided invalid, non monotonically increasing dts to muxer in stream 0)
This is an attempt to address #25.

The general idea is to have some buffer of frames ready to be encoded (hardcoded 30 below) with their associated PTS numbers. In the video files that I have encountered, the frames are given in proper order, but their PTS is not. So, we simply add the frames to a FIFO queue and the PTS into a minheap. When we encode a frame we just obtain the smallest PTS and the head of the queue for a frame, and match those together.

This way, I do not have to specify any sort of frame rate before hand. The downside is some additional delay when trying to use this real-time, but across several clips I have found 30 frames to be sufficient, and could potentially be reduced further.

Reviewed-on: https://codeberg.org/ladaapp/lada/pulls/33
Co-authored-by: AlcoholicOverfitter <beta@cyber.anonaddy.com>
Co-committed-by: AlcoholicOverfitter <beta@cyber.anonaddy.com>